+++ src/game/tests.rs
@@ -574,6 +574,7 @@ mod damage_dealt_by;
 mod damage_history;
 mod damage_limits;
 mod damage_redirection;
+mod damaged_by_this_creature;
 mod decisions_and_effects;
 mod deep_spawn;
 mod delayed_triggers;